+++ osidb/mixins.py
@@ -16,6 +16,7 @@ from django.core.exceptions import ObjectDoesNotExist, ValidationError
 from django.db import IntegrityError, models, transaction
 from django.utils import timezone
 
+from osidb.acls import ACL
 from osidb.exceptions import DataInconsistencyException
 
 from .core import generate_acls
@@ -281,64 +282,45 @@ class ACLMixinVisibility(ComparableTextChoices):
     PUBLIC = "PUBLIC", "Public"
 
 
-class ACLMixinManager(models.Manager):
-    def get_queryset(self):
-        """define base queryset for retrieving models that uses ACLs"""
-        return (
-            super()
-            .get_queryset()
-            .annotate(
-                # annotate queryset with embargoed pseudo-attribute as it is fully based on the ACLs
-                embargoed=models.Case(
-                    models.When(
-                        acl_read=[
-                            uuid.UUID(acl)
-                            for acl in generate_acls(settings.EMBARGO_READ_GROUPS)
-                        ],
-                        then=True,
-                    ),
-                    default=False,
-                    output_field=models.BooleanField(),
-                ),
-                # annotate queryset with visibility pseudo-attribute based on ACL read groups
-                visibility=models.Case(
-                    models.When(
-                        acl_read=[
-                            uuid.UUID(acl)
-                            for acl in generate_acls(settings.EMBARGO_READ_GROUPS)
-                        ],
-                        then=models.Value(ACLMixinVisibility.EMBARGOED),
-                    ),
-                    models.When(
-                        acl_read=[
-                            uuid.UUID(acl)
-                            for acl in generate_acls(settings.INTERNAL_READ_GROUPS)
-                        ],
-                        then=models.Value(ACLMixinVisibility.INTERNAL),
-                    ),
-                    models.When(
-                        acl_read=[
-                            uuid.UUID(acl)
-                            for acl in generate_acls(settings.PUBLIC_READ_GROUPS)
-                        ],
-                        then=models.Value(ACLMixinVisibility.PUBLIC),
-                    ),
-                    default=models.Value(ACLMixinVisibility.PUBLIC),
-                    output_field=models.CharField(),
-                ),
-            )
-        )
-
-
 class ACLMixin(models.Model):
     """
     mixin for models requiring access controls
     defining necessary attributes and validations
     """
 
-    objects = ACLMixinManager()
     acl_read = fields.ArrayField(models.UUIDField(), default=list)
     acl_write = fields.ArrayField(models.UUIDField(), default=list)
+
+    embargoed = models.GeneratedField(
+        expression=models.Case(
+            models.When(acl_read=ACL.EMBARGO.uuid_read, then=True),
+            default=False,
+            output_field=models.BooleanField(),
+        ),
+        output_field=models.BooleanField(),
+        db_persist=True,
+    )
+
+    visibility = models.GeneratedField(
+        expression=models.Case(
+            models.When(
+                acl_read=ACL.EMBARGO.uuid_read,
+                then=models.Value(ACLMixinVisibility.EMBARGOED),
+            ),
+            models.When(
+                acl_read=ACL.INTERNAL.uuid_read,
+                then=models.Value(ACLMixinVisibility.INTERNAL),
+            ),
+            models.When(
+                acl_read=ACL.PUBLIC.uuid_read,
+                then=models.Value(ACLMixinVisibility.PUBLIC),
+            ),
+            default=models.Value(ACLMixinVisibility.PUBLIC),
+            output_field=models.CharField(),
+        ),
+        output_field=models.CharField(max_length=10),
+        db_persist=True,
+    )
 
     # to be able to meaningfully print the ACL related alerts
     # we have to keep the mapping from the hashes to names
@@ -357,7 +339,9 @@ class ACLMixin(models.Model):
         self.acls_all
 
     def get_embargoed_acl():
-        return [uuid.UUID(acl) for acl in generate_acls(settings.EMBARGO_READ_GROUPS)]
+        return sorted(
+            uuid.UUID(acl) for acl in generate_acls(settings.EMBARGO_READ_GROUPS)
+        )
 
     @property
     def is_embargoed(self):
@@ -371,32 +355,30 @@ class ACLMixin(models.Model):
     def is_public(self):
         return set(self.acl_read + self.acl_write) == self.acls_public
 
-    @property
-    def visibility(self):
+    _visibility_dirty = False
+
+    def _current_visibility(self):
+        if not self._state.adding and not self._visibility_dirty:
+            return ACLMixinVisibility(self.visibility)
         if self.is_embargoed:
             return ACLMixinVisibility.EMBARGOED
         if self.is_internal:
             return ACLMixinVisibility.INTERNAL
         return ACLMixinVisibility.PUBLIC
 
-    @visibility.setter
-    def visibility(self, target):
-        """
-        Monotonically widen visibility to the given target level.
-
-        ACLMixinVisibility can only be widened (EMBARGOED -> INTERNAL -> PUBLIC),
-        never narrowed. If the current level is already at or wider than
-        the target, this is a no-op.
-        """
+    def set_visibility(self, target):
         target = ACLMixinVisibility(target)
-        if self.visibility >= target:
+        if self._current_visibility() >= target:
             return
 
-        if target == ACLMixinVisibility.PUBLIC:
-            self.set_public()
+        if target == ACLMixinVisibility.EMBARGOED:
+            self.set_embargoed()
         elif target == ACLMixinVisibility.INTERNAL:
             self.set_internal()
+        elif target == ACLMixinVisibility.PUBLIC:
+            self.set_public()
 
+        self._visibility_dirty = True
         self.set_acls_nested()
         self.set_acls_history()
 
@@ -433,8 +415,9 @@ class ACLMixin(models.Model):
             >>> acls == my_flaw.acl_read
             ... True
         """
-        acls = [self.group2acl(group) for group in groups]
+        acls = sorted(self.group2acl(group) for group in groups)
         self.acl_read = acls
+        self._visibility_dirty = True
         return acls
 
     def set_acl_write(self, *groups):
@@ -454,8 +437,9 @@ class ACLMixin(models.Model):
             >>> acls == my_flaw.acl_write
             ... True
         """
-        acls = [self.group2acl(group) for group in groups]
+        acls = sorted(self.group2acl(group) for group in groups)
         self.acl_write = acls
+        self._visibility_dirty = True
         return acls
 
     def set_public(self):
@@ -477,8 +461,6 @@ class ACLMixin(models.Model):
 
         self.set_acl_read(*settings.PUBLIC_READ_GROUPS)
         self.set_acl_write(*settings.PUBLIC_WRITE_GROUPS)
-        # Update the embargoed annotation to reflect the new ACL state
-        self.embargoed = False
 
     def set_embargoed(self):
         """
@@ -498,8 +480,6 @@ class ACLMixin(models.Model):
         """
         self.set_acl_read(*settings.EMBARGO_READ_GROUPS)
         self.set_acl_write(*settings.EMBARGO_WRITE_GROUPS)
-        # Update the embargoed annotation to reflect the new ACL state
-        self.embargoed = True
 
     def set_internal(self):
         """
@@ -519,8 +499,6 @@ class ACLMixin(models.Model):
         """
         self.set_acl_read(*settings.INTERNAL_READ_GROUPS)
         self.set_acl_write(*settings.INTERNAL_WRITE_GROUPS)
-        # Update the embargoed annotation to reflect the new ACL state
-        self.embargoed = False
 
     @cached_property
     def acls_public_read(self):
@@ -876,7 +854,7 @@ class ACLMixin(models.Model):
                     )
 
 
-class AlertManager(ACLMixinManager):
+class AlertManager(models.Manager):
     """Alert manager"""
 
     def get_queryset(self):
@@ -1101,7 +1079,7 @@ class AlertMixin(ValidateMixin):
             exclude = set()
 
         for f in self._meta.fields:
-            if f.name in exclude:
+            if f.name in exclude or f.generated:
                 continue
 
             # ValidationError may be raised here