+++ osidb/models/affect.py
@@ -20,7 +20,6 @@ from apps.exploits.mixins import AffectExploitExtensionMixin
 from apps.exploits.query_sets import AffectQuerySetExploitExtension
 from osidb.mixins import (
     ACLMixin,
-    ACLMixinManager,
     AlertMixin,
     NullStrFieldsMixin,
     TrackingMixin,
@@ -60,7 +59,7 @@ class NotAffectedJustification(models.TextChoices):
     VULN_CODE_NOT_PRESENT = "Vulnerable Code not Present"
 
 
-class AffectManager(ACLMixinManager, TrackingMixinManager):
+class AffectManager(TrackingMixinManager):
     """affect manager"""
 
     def get_queryset(self):
@@ -115,7 +114,7 @@ class AffectManager(ACLMixinManager, TrackingMixinManager):
         # If search has no results, this will now return an empty queryset
 
 
-class AffectV1Manager(ACLMixinManager, TrackingMixinManager):
+class AffectV1Manager(TrackingMixinManager):
     """affect v1 manager"""
 
     def get_queryset(self):
@@ -1018,6 +1017,7 @@ class AffectV1(AffectExploitExtensionMixin):
     updated_dt = models.DateTimeField()
     acl_read = fields.ArrayField(models.UUIDField(), default=list)
     acl_write = fields.ArrayField(models.UUIDField(), default=list)
+    embargoed = models.BooleanField()
 
     # Map to the array_agg SQL field that have all the related trackers/cvss scores
     # from the v2 affects
@@ -1203,7 +1203,7 @@ class AffectV1(AffectExploitExtensionMixin):
         return set(self.acl_read + self.acl_write) == self.acls_public
 
 
-class AffectCVSSManager(ACLMixinManager, TrackingMixinManager):
+class AffectCVSSManager(TrackingMixinManager):
     @staticmethod
     def create_cvss(affect, issuer, version, **extra_fields):
         """return a new CVSS or update an existing CVSS without saving"""